Refactor ApiModal into separate component
